+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/EmptySoftware/EmptySoftware.tsx
