+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/EmptySoftware/_styles.scss
