+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/EmptySoftware/index.ts
