+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareTab.tsx
