+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareTableConfig.tsx
