+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareVulnCount/SoftwareVulnCount.tsx
