+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareVulnCount/_styles.scss
