+++ frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareVulnCount/index.ts
